requirements.txt
# Core NLP / Translation / MT Eval
unbabel-comet==2.2.7 
fastapi==0.116.0
huggingface_hub==0.34.4
pytorch-lightning==2.5.5
torchmetrics==0.10.3 
nltk==3.9.1
numba==0.61.0
pandas==2.3.1
psutil==7.0.0
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.4
sacrebleu[ja]==2.5.1
tqdm==4.67.1
uvicorn==0.35.0
transformers==4.56.1
prisma==0.15.0
python-multipart==0.0.20

# Multimodal Support
easyocr==1.7.1
openai-whisper==20231117
torch>=2.6.0
Pillow==10.3.0
pytesseract

# Japanese OCR and NLP
manga-ocr
fugashi
unidic-lite

# File handling
pdfplumber
langdetect==1.0.9

# Transcreation via Gemini
google-genai>=1.0.0
pyyaml>=6.0